Consolidate input, stack, and error handling into utils.c
