--- src/error.c
@@ -1,19 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   error.c                                            :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mkazuhik <mkazuhik@student.42tokyo.jp>     +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/08/08 01:27:36 by mkazuhik          #+#    #+#             */
-/*   Updated: 2025/08/08 12:25:57 by mkazuhik          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "../push_swap.h"
-
-void	err_exit(void)
-{
-	ft_putstr_fd("Error\n", 2);
-	exit(1);
-}
